chore: remove duplicated type
