+++ src/shared/config/importConfig.ts
@@ -1,7 +1,23 @@
 import type { WabaConfigType } from '../types/config';
 import { WabaConfigEnum } from '../types/enums';
-import { WhatsAppConfig } from '../types/whatsapp';
 import Logger from '../utils/logger';
+
+export type WhatsAppConfig = {
+    accessToken: string;
+    appId?: string;
+    appSecret?: string;
+    phoneNumberId?: number;
+    businessAcctId?: string;
+    apiVersion?: string;
+    webhookEndpoint?: string;
+    webhookVerificationToken?: string;
+    listenerPort?: number;
+    debug?: boolean;
+    maxRetriesAfterWait?: number;
+    requestTimeout?: number;
+    privatePem?: string;
+    passphrase?: string;
+};
 
 const LIB_NAME = 'UTILS';
 const LOG_LOCAL = false;
