+++ src/shared/utils/index.ts
@@ -1,5 +1,4 @@
 export * from './buildFieldsQueryString';
-export * from './filters';
 export * from './flowTypeGuards';
 export * from './isMetaError';
 export * from './logger';
